tests: move out of sparsecoding directory
