+++ tests/inference/test_LSM.py
@@ -3,9 +3,7 @@ import unittest
 
 from sparsecoding import inference
 from tests.testing_utilities import TestCase
-from tests.sparsecoding.inference.common import (
-    DATAS, DATASET_SIZE, DATASET, DICTIONARY, PATCH_SIZE
-)
+from tests.inference.common import DATAS, DATASET, DICTIONARY
 
 
 class TestLSM(TestCase):
