+++ tests/priors/test_spike_slab.py
